+++ ns1/resource_record.go
@@ -20,23 +20,29 @@ import (
 var recordTypeStringEnum = NewStringEnum([]string{
 	"A",
 	"AAAA",
-	"ALIAS",
 	"AFSDB",
+	"ALIAS",
 	"CAA",
+	"CERT",
 	"CNAME",
+	"CSYNC",
+	"DHCID",
 	"DNAME",
 	"DS",
 	"HINFO",
+	"HTTPS",
 	"MX",
 	"NAPTR",
 	"NS",
 	"PTR",
 	"RP",
+	"SMIMEA",
 	"SPF",
 	"SRV",
+	"SVCB",
+	"TLSA",
 	"TXT",
 	"URLFWD",
-	"strings",
 })
 
 func recordResource() *schema.Resource {
@@ -228,8 +234,8 @@ func recordToResourceData(d *schema.ResourceData, r *dns.Record) error {
 	d.Set("ttl", r.TTL)
 
 	d.Set("override_ttl", nil)
-	if r.Type == "ALIAS" && r.Override_TTL != nil {
-		err := d.Set("override_ttl", *r.Override_TTL)
+	if r.Type == "ALIAS" && r.OverrideTTL != nil {
+		err := d.Set("override_ttl", *r.OverrideTTL)
 
 		if err != nil {
 			return fmt.Errorf("[DEBUG] Error setting override_ttl for: %s, error: %#v", r.Domain, err)
@@ -399,7 +405,7 @@ func resourceDataToRecord(r *dns.Record, d *schema.ResourceData) error {
 
 	if r.Type == "ALIAS" {
 		Override_TTL := d.Get("override_ttl").(bool)
-		r.Override_TTL = &Override_TTL
+		r.OverrideTTL = &Override_TTL
 	}
 
 	if v, ok := d.GetOk("link"); ok {